pyproject.toml
[project]
name = "a2a-hello-world"
version = "0.1.0"
description = "A minimal implementation of the Agent-to-Agent (A2A) protocol."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "a2a-sdk",
    "uvicorn",
    "starlette",
    "httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["a2a_server_agent", "a2a_client"]
